refactoring __init__.py and adding request body data model
