+++ api/Pipfile
@@ -18,6 +18,7 @@ requests = "*"
 gunicorn = "*"
 fastapi = "*"
 uvicorn = "*"
+pydantic = "*"
 
 [requires]
 python_version = "3.7"
